+++ packages/middleware/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @boostkit/middleware
 
+## 0.0.6
+
+### Patch Changes
+
+- @boostkit/cache@0.0.5
+
+## 0.0.5
+
+### Patch Changes
+
+- @boostkit/cache@0.0.4
+
 ## 0.0.4
 
 ### Patch Changes