+++ packages/session/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @boostkit/session
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @boostkit/core@0.0.6
+
+## 0.0.3
+
+### Patch Changes
+
+- @boostkit/core@0.0.5
+
 ## 0.0.2
 
 ### Patch Changes